Set data type

+++ nimare/meta/ibma.py
@@ -206,14 +206,17 @@ class Fishers(IBMAEstimator):
 
             z_map = _boolean_unmask(est_summary.z.squeeze(), self.inputs_["aggressive_mask"])
             p_map = _boolean_unmask(est_summary.p.squeeze(), self.inputs_["aggressive_mask"])
-            dof_map = np.tile(self.inputs_["z_maps"].shape[0] - 1, self.inputs_["z_maps"].shape[1])
+            dof_map = np.tile(
+                self.inputs_["z_maps"].shape[0] - 1,
+                self.inputs_["z_maps"].shape[1],
+            ).astype(np.int32)
             dof_map = _boolean_unmask(dof_map, self.inputs_["aggressive_mask"])
 
         else:
             n_total_voxels = self.inputs_["z_maps"].shape[1]
-            z_map = np.zeros(n_total_voxels)
-            p_map = np.zeros(n_total_voxels)
-            dof_map = np.zeros(n_total_voxels)
+            z_map = np.zeros(n_total_voxels, dtype=float)
+            p_map = np.zeros(n_total_voxels, dtype=float)
+            dof_map = np.zeros(n_total_voxels, dtype=np.int32)
             for bag in self.inputs_["data_bags"]["z_maps"]:
                 pymare_dset = pymare.Dataset(y=bag["values"])
                 est = pymare.estimators.FisherCombinationTest()
@@ -326,14 +329,16 @@ class Stouffers(IBMAEstimator):
 
             z_map = _boolean_unmask(est_summary.z.squeeze(), self.inputs_["aggressive_mask"])
             p_map = _boolean_unmask(est_summary.p.squeeze(), self.inputs_["aggressive_mask"])
-            dof_map = np.tile(self.inputs_["z_maps"].shape[0] - 1, self.inputs_["z_maps"].shape[1])
+            dof_map = np.tile(
+                self.inputs_["z_maps"].shape[0] - 1, self.inputs_["z_maps"].shape[1]
+            ).astype(np.int32)
             dof_map = _boolean_unmask(dof_map, self.inputs_["aggressive_mask"])
 
         else:
             n_total_voxels = self.inputs_["z_maps"].shape[1]
-            z_map = np.zeros(n_total_voxels)
-            p_map = np.zeros(n_total_voxels)
-            dof_map = np.zeros(n_total_voxels)
+            z_map = np.zeros(n_total_voxels, dtype=float)
+            p_map = np.zeros(n_total_voxels, dtype=float)
+            dof_map = np.zeros(n_total_voxels, dtype=np.int32)
             for bag in self.inputs_["data_bags"]["z_maps"]:
                 est = pymare.estimators.StoufferCombinationTest()
 
@@ -458,16 +463,16 @@ class WeightedLeastSquares(IBMAEstimator):
             se_map = _boolean_unmask(fe_stats["se"].squeeze(), self.inputs_["aggressive_mask"])
             dof_map = np.tile(
                 self.inputs_["beta_maps"].shape[0] - 1, self.inputs_["beta_maps"].shape[1]
-            )
+            ).astype(np.int32)
             dof_map = _boolean_unmask(dof_map, self.inputs_["aggressive_mask"])
 
         else:
             n_total_voxels = self.inputs_["beta_maps"].shape[1]
-            z_map = np.zeros(n_total_voxels)
-            p_map = np.zeros(n_total_voxels)
-            est_map = np.zeros(n_total_voxels)
-            se_map = np.zeros(n_total_voxels)
-            dof_map = np.zeros(n_total_voxels)
+            z_map = np.zeros(n_total_voxels, dtype=float)
+            p_map = np.zeros(n_total_voxels, dtype=float)
+            est_map = np.zeros(n_total_voxels, dtype=float)
+            se_map = np.zeros(n_total_voxels, dtype=float)
+            dof_map = np.zeros(n_total_voxels, dtype=np.int32)
             beta_bags = self.inputs_["data_bags"]["beta_maps"]
             varcope_bags = self.inputs_["data_bags"]["varcope_maps"]
             for beta_bag, varcope_bag in zip(beta_bags, varcope_bags):
@@ -582,17 +587,17 @@ class DerSimonianLaird(IBMAEstimator):
             tau2_map = _boolean_unmask(est_summary.tau2.squeeze(), self.inputs_["aggressive_mask"])
             dof_map = np.tile(
                 self.inputs_["beta_maps"].shape[0] - 1, self.inputs_["beta_maps"].shape[1]
-            )
+            ).astype(np.int32)
             dof_map = _boolean_unmask(dof_map, self.inputs_["aggressive_mask"])
 
         else:
             n_total_voxels = self.inputs_["beta_maps"].shape[1]
-            z_map = np.zeros(n_total_voxels)
-            p_map = np.zeros(n_total_voxels)
-            est_map = np.zeros(n_total_voxels)
-            se_map = np.zeros(n_total_voxels)
-            tau2_map = np.zeros(n_total_voxels)
-            dof_map = np.zeros(n_total_voxels)
+            z_map = np.zeros(n_total_voxels, dtype=float)
+            p_map = np.zeros(n_total_voxels, dtype=float)
+            est_map = np.zeros(n_total_voxels, dtype=float)
+            se_map = np.zeros(n_total_voxels, dtype=float)
+            tau2_map = np.zeros(n_total_voxels, dtype=float)
+            dof_map = np.zeros(n_total_voxels, dtype=np.int32)
             beta_bags = self.inputs_["data_bags"]["beta_maps"]
             varcope_bags = self.inputs_["data_bags"]["varcope_maps"]
             for beta_bag, varcope_bag in zip(beta_bags, varcope_bags):
@@ -710,17 +715,17 @@ class Hedges(IBMAEstimator):
             tau2_map = _boolean_unmask(est_summary.tau2.squeeze(), self.inputs_["aggressive_mask"])
             dof_map = np.tile(
                 self.inputs_["beta_maps"].shape[0] - 1, self.inputs_["beta_maps"].shape[1]
-            )
+            ).astype(np.int32)
             dof_map = _boolean_unmask(dof_map, self.inputs_["aggressive_mask"])
 
         else:
             n_total_voxels = self.inputs_["beta_maps"].shape[1]
-            z_map = np.zeros(n_total_voxels)
-            p_map = np.zeros(n_total_voxels)
-            est_map = np.zeros(n_total_voxels)
-            se_map = np.zeros(n_total_voxels)
-            tau2_map = np.zeros(n_total_voxels)
-            dof_map = np.zeros(n_total_voxels)
+            z_map = np.zeros(n_total_voxels, dtype=float)
+            p_map = np.zeros(n_total_voxels, dtype=float)
+            est_map = np.zeros(n_total_voxels, dtype=float)
+            se_map = np.zeros(n_total_voxels, dtype=float)
+            tau2_map = np.zeros(n_total_voxels, dtype=float)
+            dof_map = np.zeros(n_total_voxels, dtype=np.int32)
             beta_bags = self.inputs_["data_bags"]["beta_maps"]
             varcope_bags = self.inputs_["data_bags"]["varcope_maps"]
             for beta_bag, varcope_bag in zip(beta_bags, varcope_bags):
@@ -855,18 +860,18 @@ class SampleSizeBasedLikelihood(IBMAEstimator):
             )
             dof_map = np.tile(
                 self.inputs_["beta_maps"].shape[0] - 1, self.inputs_["beta_maps"].shape[1]
-            )
+            ).astype(np.int32)
             dof_map = _boolean_unmask(dof_map, self.inputs_["aggressive_mask"])
 
         else:
             n_total_voxels = self.inputs_["beta_maps"].shape[1]
-            z_map = np.zeros(n_total_voxels)
-            p_map = np.zeros(n_total_voxels)
-            est_map = np.zeros(n_total_voxels)
-            se_map = np.zeros(n_total_voxels)
-            tau2_map = np.zeros(n_total_voxels)
-            sigma2_map = np.zeros(n_total_voxels)
-            dof_map = np.zeros(n_total_voxels)
+            z_map = np.zeros(n_total_voxels, dtype=float)
+            p_map = np.zeros(n_total_voxels, dtype=float)
+            est_map = np.zeros(n_total_voxels, dtype=float)
+            se_map = np.zeros(n_total_voxels, dtype=float)
+            tau2_map = np.zeros(n_total_voxels, dtype=float)
+            sigma2_map = np.zeros(n_total_voxels, dtype=float)
+            dof_map = np.zeros(n_total_voxels, dtype=np.int32)
             for bag in self.inputs_["data_bags"]["beta_maps"]:
                 study_mask = bag["study_mask"]
                 sample_sizes_ex = [self.inputs_["sample_sizes"][study] for study in study_mask]
@@ -1014,16 +1019,16 @@ class VarianceBasedLikelihood(IBMAEstimator):
             tau2_map = _boolean_unmask(est_summary.tau2.squeeze(), self.inputs_["aggressive_mask"])
             dof_map = np.tile(
                 self.inputs_["beta_maps"].shape[0] - 1, self.inputs_["beta_maps"].shape[1]
-            )
+            ).astype(np.int32)
             dof_map = _boolean_unmask(dof_map, self.inputs_["aggressive_mask"])
         else:
             n_total_voxels = self.inputs_["beta_maps"].shape[1]
-            z_map = np.zeros(n_total_voxels)
-            p_map = np.zeros(n_total_voxels)
-            est_map = np.zeros(n_total_voxels)
-            se_map = np.zeros(n_total_voxels)
-            tau2_map = np.zeros(n_total_voxels)
-            dof_map = np.zeros(n_total_voxels)
+            z_map = np.zeros(n_total_voxels, dtype=float)
+            p_map = np.zeros(n_total_voxels, dtype=float)
+            est_map = np.zeros(n_total_voxels, dtype=float)
+            se_map = np.zeros(n_total_voxels, dtype=float)
+            tau2_map = np.zeros(n_total_voxels, dtype=float)
+            dof_map = np.zeros(n_total_voxels, dtype=np.int32)
             beta_bags = self.inputs_["data_bags"]["beta_maps"]
             varcope_bags = self.inputs_["data_bags"]["varcope_maps"]
             for beta_bag, varcope_bag in zip(beta_bags, varcope_bags):
@@ -1149,14 +1154,14 @@ class PermutedOLS(IBMAEstimator):
 
             t_map = _boolean_unmask(t_map.squeeze(), self.inputs_["aggressive_mask"])
             z_map = _boolean_unmask(z_map.squeeze(), self.inputs_["aggressive_mask"])
-            dof_map = np.tile(dof, self.inputs_["beta_maps"].shape[1])
+            dof_map = np.tile(dof, self.inputs_["beta_maps"].shape[1]).astype(np.int32)
             dof_map = _boolean_unmask(dof_map, self.inputs_["aggressive_mask"])
 
         else:
             n_total_voxels = self.inputs_["beta_maps"].shape[1]
-            t_map = np.zeros(n_total_voxels)
-            z_map = np.zeros(n_total_voxels)
-            dof_map = np.zeros(n_total_voxels)
+            t_map = np.zeros(n_total_voxels, dtype=float)
+            z_map = np.zeros(n_total_voxels, dtype=float)
+            dof_map = np.zeros(n_total_voxels, dtype=np.int32)
             for bag in self.inputs_["data_bags"]["beta_maps"]:
                 # Use intercept as explanatory variable
                 bag["tested_vars"] = np.ones((bag["values"].shape[0], 1))
@@ -1257,8 +1262,8 @@ class PermutedOLS(IBMAEstimator):
 
         else:
             n_total_voxels = self.inputs_["beta_maps"].shape[1]
-            log_p_map = np.zeros(n_total_voxels)
-            z_map = np.zeros(n_total_voxels)
+            log_p_map = np.zeros(n_total_voxels, dtype=float)
+            z_map = np.zeros(n_total_voxels, dtype=float)
             for bag in self.inputs_["data_bags"]["beta_maps"]:
                 log_p_map_tmp, t_map_tmp, _ = permuted_ols(
                     bag["tested_vars"],
